refactor featurization module
