+++ chemprop/v2/featurizers/__init__.py
@@ -1,6 +1,4 @@
-from .protos import *
-from .atom import AtomFeaturizer
-from .bond import BondFeaturizer
-from .molgraph import MolGraph
-from .molecule import MoleculeMolGraphFeaturizer
-from .reaction import CGRFeaturizer, RxnMode
+from .atom import AtomFeaturizer, AtomFeaturizerProto
+from .bond import BondFeaturizer, BondFeaturizerProto
+from .molgraph import *
+from .molecule import *
