+++ chemprop/v2/tests/unit/data/test_datapoint.py
@@ -3,7 +3,7 @@ import pytest
 
 from chemprop.v2.data import MoleculeDatapoint
 #from chemprop.featurizers.features_generators import get_available_features_generators
-from chemprop.v2.featurizers.featurizers import MoleculeFeaturizerRegistry
+from chemprop.v2.featurizers.molecule import MoleculeFeaturizerRegistry
 
 @pytest.fixture(
     params=[
